Add unit tests for ChatbotRouter and refactor test structure for improved clarity
